Add a function to get the first responder.
This CL adds a function and a category on UIResponder to get to the
first responder.
Based on http://iosdeveloperzone.com/2013/05/20/snippet-finding-the-current-first-responder/,
it leverages the UIApplication's -sendAction:toObject:nil to message
the first responder and make it mark itself as the first responder.

BUG=581289
R=rohitrao@chromium.org

Review URL: https://codereview.chromium.org/1639783002

Cr-Original-Commit-Position: refs/heads/master@{#372338}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 319093bc7667e64f0adb8d709975486f4169be5a