Change ownership of BookmarkClient.
BookmarkModel now owns the BookmarkClient (there is no real need for
BookmarkClient to be a KeyedService). As was found in other components
having the client be owned by the service reduce the risk of breaking
the componentization.

BUG=None

Review URL: https://codereview.chromium.org/1586013002

Cr-Original-Commit-Position: refs/heads/master@{#372654}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 070a510da9d775ac7b31dc6bfd5e37a7fe9153c8