Adds the OCHamcrest library to the iOS build.
OCHamcrest is a dependency of the integration testing framework that will
be used by iOS code. It is only used for tests (and is marked as so in
ios/third_party/ochamcrest/BUILD.gn).

The build is a bit involved as OCHamcrest as it mix relative imports "..."
with imports of the form <OCHamcrest/...> but the source code is not found
in an OCHamcrest directory. So build such a directory structure in
<(SHARED_INTERMEDIATE_DIR)/ios/third_party/ochamcrest as part of the build.

TBR=dpranke@chromium.org
BUG=580730
TEST=None

Review URL: https://codereview.chromium.org/1723663002

Cr-Original-Commit-Position: refs/heads/master@{#376985}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a9045f46f887e536b148ed436851b132d8e1025c