CREDENTIAL: Only toggle 'skip_zero_click' state if API is used.
Currently, we're toggling 'skip_zero_click' to 'false' if the user logs
in successfully after autofill. We should restrict that a bit to ensure
that the site understands the credential management API (and can therefore
be assumed to explain to us when the user signs out).

This patch adjusts the API's behavior to send the relevant PasswordForm
to the client whenever it could have been handed over but wasn't. Upon
successful autofill-based login with that same credential, the client is
responsible for toggling its 'skip_zero_click' value, and for popping up
the first-run experience if necessary.

BUG=587440
R=vasilii@chromium.org,vabr@chromium.org

Review URL: https://codereview.chromium.org/1735013004

Cr-Original-Commit-Position: refs/heads/master@{#378190}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ce3eee6988a1d901f47f95cf192b77874b85bdb2