Use password_manager::metrics_util::ResponseType on iOS
The passwords infobar delegate had a separate copy of the ResponseType enum. So
far it had a subset of values of the desktop version of ResponseType (matching
the positions), so it was not a problem, but to avoid diverging in the future,
this CL switches the iOS code to use the component-defined ResponseType (which
also matches histograms.xml).

R=melandory@chromium.org

BUG=593295

Review URL: https://codereview.chromium.org/1777473004

Cr-Original-Commit-Position: refs/heads/master@{#380132}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0090ffb1c2878f85bd94750db07c09448741172b