Add a unittest for PasswordController
While PasswordController already has a unittest downstream, that one is rather
an integration test (the heaviness of which is the reason it is still
downstream, stuck on some downstream-only testing framework).

When a test was needed for https://codereview.chromium.org/1806333005/, it
turned out a simple unittest would be better. This CL adds that test and
places it upstream. (Ultimately we should convert the downstream tests to a
lighter version and add them here as well, but not in this CL.)

BUG=595717, 598672

Review URL: https://codereview.chromium.org/1808853005

Cr-Original-Commit-Position: refs/heads/master@{#383941}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 86b472697e7f5871d75963d32266332b1a95a242