[ios] Added WebState::GetLoadingProgress web// API.
This API is necessary to remove loading progress code from Tab, which is
currently stores the loading progress and provides it to Toolbar.
Corresponding method does not exist in WebContents, because on Desktop
the progress is not shown in UI and on Android progress is stored in
WebContentsDelegateAndroid.

BUG=None

Review URL: https://codereview.chromium.org/1905393002

Cr-Original-Commit-Position: refs/heads/master@{#389887}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec7db8a2d1757e5d99200d12147d5a3e77d474e0