Support URL-based deletion for Channel IDs.
ChannelID::server_identifier() contains a hostname. We convert it to
a corresponding HTTPS origin, and match it with the provided URL.

For example, a Channel ID for the host "google.com" will be converted
to GURL("https://google.com"). This Channel ID will then be matched by:

- a deletion request from RegistrableDomainFilterBuilder for "google.com",
  because the url "https://google.com" is matched by the host "google.com"
- a deletion request from OriginFilterBuilder for "https://google.com",
  because url "https://google.com" is trivially matched by
  url::Origin("https://google.com")

BUG=589586

Review-Url: https://codereview.chromium.org/1913853002
Cr-Original-Commit-Position: refs/heads/master@{#390932}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e9adacbf6d9a0aa894e7eb998a82a1932bab4c93