Add histograms for notices about other forms of browsing history
Notices about the existence of other forms of browsing history on
history.google.com are shown to users in the header of the History page,
footer of the Clear Browsing Data dialog, and an additional dialog on top
of the CBD dialog after history deletion; but only to those users for
whom they are relevant.

This CL adds histograms that are recorded whenever one of the described
UI elements is shown and records whether we also showed the notice or not
(thus computing the amount and ratio of users for whom we decided to show
the notice).

Implemented:
- History page (shared between Desktop and Android)
- History page (iOS)
- CBD dialog (Desktop WebUI)
- CBD dialog (Desktop MD, only the footer)
- CBD dialog (Android)

Missing (will be added in followup CLs):
- CBD dialog (iOS) - not part of Chromium repo
- CBD dialog (Desktop MD, history deletion dialog) - does not exist yet

BUG=608842

Review-Url: https://codereview.chromium.org/1948633002
Cr-Original-Commit-Position: refs/heads/master@{#391682}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f22966d3e45226f3048405f9f5d7d6898285d76