[NTP Snippets] Refactor JSON-parsing to fetcher
Advantages include:
1. Removing complexity from NTPSnippetsService.
2. Centralizing more error-cases in the fetcher.
3. Improved unit-tests (NTPSnippetsServiceTest now exercises public
   API).

In this first step, the fetcher's API is changed to expose a base::Value
containing pre-parsed JSON, and representing a dictionary that contains
snippets. Follow-up patches will replace base::Value with instances of
NTPSnippet.

The callback to parse JSON now needs to be injected to the fetcher
instead of NTPSnippetsService.

BUG=584428

Review-Url: https://codereview.chromium.org/1947043002
Cr-Original-Commit-Position: refs/heads/master@{#392253}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2c3a80ff0f920eb5a87d1d3f80ad429f46e6a8fc