[Cleanup] Removed unused debugging code.
The debugging CHECK removed here has been added in 2011, with a final
comment of "More CHECKs are needed" on the bug, and no further updates
since then. Looking at crash info, this does not seem to trigger in the
wild. (Or if it does, at miniscule rates)

BUG=102004

Review-Url: https://codereview.chromium.org/1981273004
Cr-Original-Commit-Position: refs/heads/master@{#394623}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 606a2ea600a2f7633db368a953a14eb30f0be59f