Remove MessageLoop::current()->task_runner() in ios.
This CL replaces MessageLoop::current()->task_runner() with
ThreadTaskRunnerHandle::Get() in ios.

In files where this replacement is made, it adds these include:
  #include "base/threading/thread_task_runner_handle.h"
  #include "base/location.h" (if required)
  #include "base/single_thread_task_runner.h" (if required)
And removes this include if it is no longer required:
  #include "base/message_loop/message_loop.h"

Why?
The fact that there's a MessageLoop on the thread is an
unnecessary implementation detail. When browser threads
are migrated to base/task_scheduler, tasks will no longer
have access to a MessageLoop but they will have access to
a ThreadTaskRunnerHandle.

Steps to generate this patch:
1. Run message_loop_cleanup_2.py (see code on the bug).
2. Run tools/sort-headers.py on modified files.
3. Run git cl format.

BUG=616447
R=droger@chromium.org

Review-Url: https://codereview.chromium.org/2054563002
Cr-Original-Commit-Position: refs/heads/master@{#398863}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf973b866da1b91f88b9a4e5e18daaf55a831e08