Add migration from old to new NTP Promo pref structure.
The structure of the pref that stores NTP Promos changed when support
was added for storing multiple NTP Promos. This CL adds code to handle
migrating the data from the old structure to the new structure, so that
there is no data loss when users upgrade.

BUG=620554

Review-Url: https://codereview.chromium.org/2094523002
Cr-Original-Commit-Position: refs/heads/master@{#402356}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d7ff2e8d36006998c5eadbc5f560990de231dcd