SelectorCoordinator for presenting a picker view from an infobar
Add coordinator for presenting a picker view from an infobar, as needed
in update passwords where there are multiple accounts, or in the
translate infobar for language selection. Starting the coordinator
animates the picker view in from the bottom edge of the
baseViewController's view, and stopping animates it out.

BUG=622244

Review-Url: https://codereview.chromium.org/2105253004
Cr-Original-Commit-Position: refs/heads/master@{#403780}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b7a60b6c46e6a3c703a714d64b1786ad130c76c8