UpdatePasswordInfoBarController for update password UI
Adds UpdatePasswordInfoBarController as infobar controller for the
update passwords infobar. UpdatePasswordInfoBarController subclasses
ConfirmInfoBarController, adding an additional link to the label for
choosing an account in the case where multiple accounts are possible for
the password update. Tapping on this link presents a picker view for
selecting the appropriate account.

BUG=622244

Review-Url: https://codereview.chromium.org/2106413003
Cr-Original-Commit-Position: refs/heads/master@{#403872}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 16245bede16eea10edd85949b1da70c4e033f798