Componentize spellcheck [1]: create component and move switches there.
Current spellcheck code lives in the chrome/ layer, but we want to make it
available in webview as well. Therefore this effort of making a dedicated
spellcheck component. As a first step we create a //components/spellcheck
and move the chrome spellcheck related switches there.

BUG=583616, 629609

Review-Url: https://codereview.chromium.org/2166683003
Cr-Original-Commit-Position: refs/heads/master@{#407469}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 677f301fb8de766a62d065408821722bb89f85f3