Create IOSChromeContentSuggestionsServiceFactory
Create the factory so that the new ContentSuggestionsService is
available on iOS. It is currently always disabled, but the factory
ensures that the service doesn't introduce new non-iOS dependencies.

Remove the IOSChromeNTPSnippetsServiceFactory because the
NTPSnippetsService is now a plain provider and will not be
a KeyedService anymore. It will be created by the new
IOSChromeContentSuggestionsServiceFactory in the future.

BUG=632690
Owner: me

Review-Url: https://codereview.chromium.org/2194793004
Cr-Original-Commit-Position: refs/heads/master@{#408934}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ca76f9c6a37b34de6d70b3e4e1ddb9c6f61c3041