Mechanical change in the component updater Configurator.
This is the first CL in a series of CLs to implement component updater
group policies

The CL renames a few functions in the Configurator, for consistency.
It introduces Configurator::EnabledComponentUpdates and
CrxComponent::supports_group_policy_enable_component_updates.

These members are placeholders and have no effect on the current
code behavior.

BUG=624128

Review-Url: https://codereview.chromium.org/2199423002
Cr-Original-Commit-Position: refs/heads/master@{#409337}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb4e5e933ae9975ccc24e83f48369716d65b669d