RetryableURLFetcher needs to handle invalid URL as a permanent failure.
RetryableURLFetcher relies on a -urlToFetch delegate method. It used to
treat invalid URL returned as a temporary failure. This change now
treats it as a permanent failure by calling the response handler method
with nil to indicate so.

Without this change, user of this RetryableURLFetcher may initiate a
retry at the next cold start.

BUG=632202

Review-Url: https://codereview.chromium.org/2208203002
Cr-Original-Commit-Position: refs/heads/master@{#409770}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 50cd6e2fad8dd26cb0362297440549feffa1ebbd