Expose Physical Web data
The goal is to create an API for pushing changes of Physical Web data,
such as new URLs found, so that other parties can react to the change
without constantly pulling the data. To do this we are providing an API
that allows others to register a listener for push updates when data
changes.

BUG=636490

Review-Url: https://codereview.chromium.org/2231983002
Cr-Original-Commit-Position: refs/heads/master@{#413948}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1098aecd3818cf9eef26cfa77aed4ae51f6621c6