Make PasswordFormManager::best_matches_ const
...and also the related lists of passwords from PasswordStore.
PasswordFormManager should not really modify them, and will not be able to do
so once their ownership will transfer to the coming FormFetcher, shared among
multiple PasswordFormManagers.

As a side-effect, this CL also gets rid of two aliases for maps involving
PasswordForm. While the map's signature is long, it is more helpful to spell it
out than to hide it behind a cryptic PasswordFormMap, or similar, which does
not hint on ownership involved in the map.

BUG=621355

Review-Url: https://codereview.chromium.org/2262843002
Cr-Original-Commit-Position: refs/heads/master@{#414016}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5406f79261a94c55b0826fac05f73375ded7d15