Verify that the content offset is not reset when frame is changed.
When the web view frame is changed, the content offset can be reset,
causing the toolbar to cover the top of PDFs. This CL ensures that
if the content offset is changed when setting the frame, it is restored
to the correct value.

BUG=641323

Review-Url: https://codereview.chromium.org/2324703002
Cr-Original-Commit-Position: refs/heads/master@{#417884}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe5fb8dd0a77a43120793e60bd20f6d668f1e1c1