Do not flush SequencedWorkerPools in browser tests.
Before this CL, the blocking pool and the simple cache's pool were
flushed at the end of each browser test. That means that every task
posted to these pools ran, regardless of their shutdown behavior.
This flush was introduced by https://codereview.chromium.org/11649032
to prevent leaks between tests (see https://crbug.com/168415).

This flush is not required for browser tests since they each run in
their own process.

Note that in production, BrowserThreads are joined before the
blocking pool. Before this CL, the blocking pool was flushed before
BrowserThreads were joined. With this CL, the behavior of browser
threads is closer to production.

Why is this important? There is no way to flush TaskScheduler and
we would prefer not to implement it if it is not required.

BUG=622400

Review-Url: https://codereview.chromium.org/2355093003
Cr-Original-Commit-Position: refs/heads/master@{#420366}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0881867ce08db7368ca9f6ea10aa230dc2849227