Add a cache counter for iOS.
The CacheCounter class in chrome/browser/browsing_data could not be
componentized (unlike other counters in components/browsing_data) because
of its usage of StoragePartitionHttpCacheDataRemover that has a dependency
on content/.

This CL implements a separate CacheCounter class for iOS with the same
functionality. The core of the implementation, a series of IO thread
callbacks, is delegated to the private IOThreadCacheCounter class that
mimics StoragePartitionHttpCacheDataRemover.

BUG=620317

Review-Url: https://codereview.chromium.org/2354643002
Cr-Original-Commit-Position: refs/heads/master@{#420599}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2d290dc4a1bcac7ad6cd67553e5831fa3680d6bb