[Sync] Move //components/sync to the syncer namespace.
Things in components should live in namespaces that match their
component. Unfortunately "sync" is a symbol defined in the std library,
so in this case we are using "syncer".

This CL does the following things inside //components/sync:

- Replaces all uses of syncer_v2, sync_driver, sync_driver_v2, and browser_sync with syncer.
- Adds syncer namespace to some classes that were missing a namespace.
- Removes almost all "syncer::" prefixes since they became redundant.

Everything outside the component is just fixing namespace prefixes and forward declarations.
I also fixed *some* lints, especially in headers outside of the component.

This script was the starting point for this CL: http://pastebin.com/PPhPTtRK

BUG=646685
TBR=caitkp

Review-Url: https://codereview.chromium.org/2376123003
Cr-Original-Commit-Position: refs/heads/master@{#422239}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a1a854356c262a2e501dc3f30ea93585f1c8635a